Extract Source::with_annotation
